+++ src/untangled/openid_client.cljs
@@ -1,9 +1,18 @@
 (ns untangled.openid-client
   (:require [clojure.string :as s]
-            [clojure.walk :as w]))
+            [clojure.walk :as w])
+  (:import goog.net.Cookies))
 
 (defn params []
   (apply str (rest (-> js/window .-location .-hash))))
+
+(defn get-tokens-from-cookies []
+  (let [cookies     (Cookies. js/document)
+        cookie-keys (.getKeys cookies)]
+    (reduce
+      #(assoc %1 %2 (.get cookies %2))
+      {}
+      cookie-keys)))
 
 (defn tokens-from-params [params]
   (apply merge
@@ -21,7 +30,9 @@
   Also composes in a request-transform in networking's send function
   to add an Authorization header for each request."
   [app]
-  (let [tokens (tokens-from-params (params))
+  (let [tokens (or
+                 (get-tokens-from-cookies)
+                 (tokens-from-params (params)))
         id-claims (-> tokens (get "id_token") parse-claims js->clj w/keywordize-keys)]
     (swap! app update-in [:initial-state]
            (fn [m]